Rey_Personal_Final/backend/dog/urls.py
"""dog URL Configuration

The `urlpatterns` list routes URLs to views. 
Example:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from rest_framework_simplejwt.views import TokenRefreshView,TokenObtainPairView
from account.api import RegisterApi
from account.api import CurrentInfoApi
from order.views import OrderViewSet

router = DefaultRouter()
# register a router group
router.register('orders', OrderViewSet)

urlpatterns = [
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/register/', RegisterApi.as_view()),
    path('api/current_user/', CurrentInfoApi.as_view()),
    path('api/', include(router.urls)),
]