Rey_Personal_Final/backend/requirements.txt
asgiref==3.6.0
Django==4.1.6
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
install==1.3.5
pytz==2022.7.1
sqlparse==0.4.3
